:sparkles: feat(product): add check for product in orders before deletion and handle errors accordingly

+++ src/domain/ecommerce/application/use-cases/delete-product.ts
@@ -2,16 +2,24 @@ import { left, right, type Either } from '@/core/either'
 import { ResourceNotFoundError } from '@/core/errors/errors/resource-not-found-error'
 import { Injectable } from '@nestjs/common'
 import { ProductsRepository } from '../repositories/products-repository'
+import { OrderItemsRepository } from '../repositories/order-items-repository'
+import { ProductInOrderError } from './errors/product-in-order-error'
 
 interface DeleteProductUseCaseRequest {
   id: string
 }
 
-type DeleteProductUseCaseResponse = Either<ResourceNotFoundError, null>
+type DeleteProductUseCaseResponse = Either<
+  ResourceNotFoundError | ProductInOrderError,
+  null
+>
 
 @Injectable()
 export class DeleteProductUseCase {
-  constructor(private readonly productsRepository: ProductsRepository) {}
+  constructor(
+    private readonly productsRepository: ProductsRepository,
+    private readonly orderItemsRepository: OrderItemsRepository
+  ) {}
 
   async execute({
     id,
@@ -20,6 +28,12 @@ export class DeleteProductUseCase {
 
     if (!product) {
       return left(new ResourceNotFoundError('Product'))
+    }
+
+    const orderItems = await this.orderItemsRepository.findManyByProductId(id)
+
+    if (orderItems.length > 0) {
+      return left(new ProductInOrderError(id))
     }
 
     await this.productsRepository.delete(product)
